I created a branch develop_feat_8.0Client(March_29_2022). It's just a proof of concept: - Made based on the develop_feat_quest&doodad branch; - Requires careful correction of package structures; - Database compact.sqlite3 from server version 3.0.3.0; - Client version 8.0.3.12 r558734 14.12.2021 ArcheAge Kakao.

+++ AAEmu.Game/Core/Managers/AccountManager.cs
@@ -11,11 +11,11 @@ namespace AAEmu.Game.Core.Managers
     {
         private static Logger _log = LogManager.GetCurrentClassLogger();
 
-        private ConcurrentDictionary<uint, GameConnection> _accounts;
+        private ConcurrentDictionary<ulong, GameConnection> _accounts;
 
         public AccountManager()
         {
-            _accounts = new ConcurrentDictionary<uint, GameConnection>();
+            _accounts = new ConcurrentDictionary<ulong, GameConnection>();
             TickManager.Instance.OnTick.Subscribe(RemoveDeadConnections, TimeSpan.FromSeconds(30));
         }
 
@@ -36,12 +36,12 @@ namespace AAEmu.Game.Core.Managers
             }
         }
 
-        public void Remove(uint id)
+        public void Remove(ulong id)
         {
             _accounts.TryRemove(id, out _);
         }
 
-        public bool Contains(uint id)
+        public bool Contains(ulong id)
         {
             return _accounts.ContainsKey(id);
         }
