+++ AAEmu.Game/Models/Game/Char/CharacterSkills.cs
@@ -37,9 +37,9 @@ namespace AAEmu.Game.Models.Game.Char
         {
             var template = SkillManager.Instance.GetSkillTemplate(skillId);
             if (template.AbilityId > 0 &&
-                template.AbilityId != (byte) Owner.Ability1 &&
-                template.AbilityId != (byte) Owner.Ability2 &&
-                template.AbilityId != (byte) Owner.Ability3)
+                template.AbilityId != (int)Owner.Ability1 &&
+                template.AbilityId != (int)Owner.Ability2 &&
+                template.AbilityId != (int)Owner.Ability3)
                 return;
             var points = ExpirienceManager.Instance.GetSkillPointsForLevel(Owner.Level);
             points -= GetUsedSkillPoints();
@@ -58,7 +58,7 @@ namespace AAEmu.Game.Models.Game.Char
             {
                 Id = template.Id,
                 Template = template,
-                Level = (template.LevelStep > 0 ? (byte)(((Owner.GetAbLevel((AbilityType)template.AbilityId) - (template.AbilityLevel)) / template.LevelStep) + 1): (byte)1)
+                Level = template.LevelStep > 0 ? (byte)((Owner.GetAbLevel((AbilityType)template.AbilityId) - template.AbilityLevel) / template.LevelStep + 1): (byte)1
             };
             Skills.Add(skill.Id, skill);
 
@@ -69,10 +69,10 @@ namespace AAEmu.Game.Models.Game.Char
         public void AddBuff(uint buffId)
         {
             var template = SkillManager.Instance.GetPassiveBuffTemplate(buffId);
-            if(template.AbilityId > 0 && 
-               template.AbilityId != (byte)Owner.Ability1 && 
-               template.AbilityId != (byte)Owner.Ability2 && 
-               template.AbilityId != (byte)Owner.Ability3)
+            if (template.AbilityId > 0 &&
+               template.AbilityId != (int)Owner.Ability1 &&
+               template.AbilityId != (int)Owner.Ability2 &&
+               template.AbilityId != (int)Owner.Ability3)
                 return;
             var points = ExpirienceManager.Instance.GetSkillPointsForLevel(Owner.Level);
             points -= GetUsedSkillPoints();
