+++ AAEmu.Game/Models/Game/Skills/Effects/BuffEffect.cs
@@ -40,7 +40,7 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
             if (target.Buffs.CheckBuffImmune(Buff.Id))
                 return; // TODO send error of immune?
 
-            uint abLevel = 1;
+            ushort abLevel = 1;
             if (caster is Character character)
             {
                 if (source.Skill != null)
@@ -48,12 +48,12 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
                     var template = source.Skill.Template;
                     var abilityLevel = character.GetAbLevel((AbilityType)source.Skill.Template.AbilityId);
                     if (template.LevelStep != 0)
-                        abLevel = (uint)((abilityLevel / template.LevelStep) * template.LevelStep);
+                        abLevel = (ushort)(abilityLevel / template.LevelStep * template.LevelStep);
                     else
-                        abLevel = (uint)template.AbilityLevel;
+                        abLevel = (ushort)template.AbilityLevel;
 
                     //Dont allow lower than minimum ablevel for skill or infinite debuffs can happen
-                    abLevel = (uint)Math.Max(template.AbilityLevel, (int)abLevel);
+                    abLevel = (ushort)Math.Max(template.AbilityLevel, abLevel);
                 }
                 else if (source.Buff != null)
                 {
@@ -64,7 +64,7 @@ namespace AAEmu.Game.Models.Game.Skills.Effects
             {
                 if(source.Skill != null)
                 {
-                    abLevel = (uint)source.Skill.Template.AbilityLevel;
+                    abLevel = (ushort)source.Skill.Template.AbilityLevel;
                 }
             }
